+++ src/audio/CMakeLists.txt
@@ -177,11 +177,13 @@ if(CONFIG_IPC_MAJOR_3)
 	set(src_sources src/src.c src/src_ipc3.c src/src_generic.c)
 	set(eq-iir_sources eq_iir/eq_iir_ipc3.c eq_iir/eq_iir_generic.c)
 	set(eq-fir_sources eq_fir/eq_fir_ipc3.c)
+	set(tdfb_sources tdfb/tdfb_ipc3.c)
 elseif(CONFIG_IPC_MAJOR_4)
 	set(volume_sources volume/volume.c volume/volume_generic.c volume/volume_ipc4.c)
 	set(src_sources src/src.c src/src_ipc4.c src/src_generic.c)
 	set(eq-iir_sources eq_iir/eq_iir_ipc4.c eq_iir/eq_iir_generic.c)
 	set(eq-fir_sources eq_fir/eq_fir_ipc4.c)
+	set(tdfb_sources tdfb/tdfb_ipc4.c)
 endif()
 set(mixer_sources ${mixer_src})
 set(asrc_sources asrc/asrc.c asrc/asrc_farrow.c asrc/asrc_farrow_generic.c)
