+++ src/audio/tdfb/CMakeLists.txt
@@ -1,3 +1,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 add_local_sources(sof tdfb.c tdfb_generic.c tdfb_hifiep.c tdfb_hifi3.c tdfb_direction.c)
+if(CONFIG_IPC_MAJOR_3)
+	add_local_sources(sof tdfb_ipc3.c)
+elseif(CONFIG_IPC_MAJOR_4)
+	add_local_sources(sof tdfb_ipc4.c)
+endif()
